# Use pointers instead of binary encoding