--- a/table/table.go
+++ b/table/table.go
@@ -17,7 +17,6 @@
 package table
 
 import (
-	"encoding/binary"
 	"fmt"
 	"io"
 	"os"
@@ -277,7 +276,7 @@ func (t *Table) readIndex() error {
 	// Read checksum len from the last 4 bytes.
 	readPos -= 4
 	buf := t.readNoFail(readPos, 4)
-	checksumLen := int(binary.BigEndian.Uint32(buf))
+	checksumLen := int(y.BytesToU32(buf))
 
 	// Read checksum.
 	expectedChk := &pb.Checksum{}
@@ -290,7 +289,7 @@ func (t *Table) readIndex() error {
 	// Read index size from the footer.
 	readPos -= 4
 	buf = t.readNoFail(readPos, 4)
-	indexLen := int(binary.BigEndian.Uint32(buf))
+	indexLen := int(y.BytesToU32(buf))
 	// Read index.
 	readPos -= indexLen
 	data := t.readNoFail(readPos, indexLen)
@@ -323,15 +322,15 @@ func (t *Table) block(idx int) (*block, error) {
 
 	// Read meta data related to block.
 	readPos := len(blk.data) - 4 // First read checksum length.
-	blk.chkLen = int(binary.BigEndian.Uint32(blk.data[readPos : readPos+4]))
+	blk.chkLen = int(y.BytesToU32(blk.data[readPos : readPos+4]))
 
 	// Skip reading checksum, and move position to read numEntries in block.
 	readPos -= (blk.chkLen + 4)
-	blk.numEntries = int(bytesToU32(blk.data[readPos : readPos+4]))
+	blk.numEntries = int(y.BytesToU32(blk.data[readPos : readPos+4]))
 	entriesIndexStart := readPos - (blk.numEntries * 4)
 	entriesIndexEnd := entriesIndexStart + blk.numEntries*4
 
-	blk.entryOffsets = bytesToU32Slice(blk.data[entriesIndexStart:entriesIndexEnd])
+	blk.entryOffsets = y.BytesToU32Slice(blk.data[entriesIndexStart:entriesIndexEnd])
 
 	blk.entriesIndexStart = entriesIndexStart
 	// Verify checksum on if checksum verification mode is OnRead on OnStartAndRead.